fix(web): satisfy eslint react-hooks and react-refresh rules
- Defer setState from effects via queueMicrotask (markdown editor, grades)
- Math popover: position from layout effect + microtask; math toolbar anchor via registerMathAnchor callback
- Split shell nav context vs useShellNav; split math parse util for refresh-only components
