+++ clients/web/src/components/layout/shell-nav-context-core.ts
@@ -1,0 +1,9 @@
+import { createContext, type Dispatch, type SetStateAction } from 'react'
+
+export type ShellNavContextValue = {
+  mobileNavOpen: boolean
+  setMobileNavOpen: Dispatch<SetStateAction<boolean>>
+  closeMobileNav: () => void
+}
+
+export const ShellNavContext = createContext<ShellNavContextValue | null>(null)
